Refactor App using classes
