+++ src/App.js
@@ -1,12 +1,15 @@
+import React from 'react';
 import './App.css';
-import Calculator from './components/calculator';
+import Calculator from './components/Calculator';
 
-function App() {
-  return (
-    <div className='App'>
-      <Calculator text="Calculator"/>
-    </div>
-  );
+class App extends React.Component {
+  render() {
+    return (
+      <div className='App'>
+        <Calculator text='Calculator' />
+      </div>
+    );
+  }
 }
 
 export default App;
